MOHAWK: LB - Cleanup ~4000 LoC in main LB (I).
Move bookinfo and UI menu related into their respective files.
Move LBAnimationNode into its own file.
Move LBPage, BITL into livingbooks_page.cpp
Create livingbooks_itemtypes.cpp to handle misc item types.
