+++ engines/mohawk/livingbooks.cpp
@@ -75,62 +75,6 @@ Common::Rect MohawkEngine_LivingBooks::readRect(Common::ReadStreamEndian *stream
 	return rect;
 }
 
-LBPage::LBPage(MohawkEngine_LivingBooks *vm) : _vm(vm) {
-	_code = nullptr;
-	_mhk = nullptr;
-
-	_baseId = 0;
-	_cascade = false;
-}
-
-void LBPage::open(Archive *mhk, uint16 baseId) {
-	_mhk = mhk;
-	_baseId = baseId;
-
-	_vm->addArchive(_mhk);
-	if (!_vm->hasResource(ID_BCOD, baseId)) {
-		// assume that BCOD is mandatory for v4/v5
-		if (_vm->getGameType() == GType_LIVINGBOOKSV4 || _vm->getGameType() == GType_LIVINGBOOKSV5)
-			error("missing BCOD resource (id %d)", baseId);
-		_code = new LBCode(_vm, 0);
-	} else {
-		_code = new LBCode(_vm, baseId);
-	}
-
-	loadBITL(baseId);
-	for (uint i = 0; i < _items.size(); i++)
-		_vm->addItem(_items[i]);
-
-	for (uint32 i = 0; i < _items.size(); i++)
-		_items[i]->init();
-
-	for (uint32 i = 0; i < _items.size(); i++)
-		_items[i]->startPhase(kLBPhaseLoad);
-}
-
-void LBPage::addClonedItem(LBItem *item) {
-	_vm->addItem(item);
-	_items.push_back(item);
-}
-
-void LBPage::itemDestroyed(LBItem *item) {
-	for (uint i = 0; i < _items.size(); i++)
-		if (item == _items[i]) {
-			_items.remove_at(i);
-			return;
-		}
-	error("itemDestroyed didn't find item");
-}
-
-LBPage::~LBPage() {
-	delete _code;
-	_vm->removeItems(_items);
-	for (uint i = 0; i < _items.size(); i++)
-		delete _items[i];
-	_vm->removeArchive(_mhk);
-	delete _mhk;
-}
-
 MohawkEngine_LivingBooks::MohawkEngine_LivingBooks(OSystem *syst, const MohawkGameDescription *gamedesc) : MohawkEngine(syst, gamedesc) {
 	DebugMan.addDebugChannel(kDebugCode, "Code", "Track Script Execution");
 
@@ -302,65 +246,6 @@ void MohawkEngine_LivingBooks::pauseEngineIntern(bool pause) {
 			_video->resumeVideos();
 		}
 		_system->updateScreen();
-	}
-}
-
-void MohawkEngine_LivingBooks::loadBookInfo(const Common::Path &filename) {
-	_bookInfoFile.allowNonEnglishCharacters();
-	if (!_bookInfoFile.loadFromFile(filename))
-		error("Could not open %s as a config file", filename.toString().c_str());
-
-	_title = getStringFromConfig("BookInfo", "title");
-	_copyright = getStringFromConfig("BookInfo", "copyright");
-
-	_numPages = getIntFromConfig("BookInfo", "nPages");
-	_numLanguages = getIntFromConfig("BookInfo", "nLanguages");
-	_screenWidth = getIntFromConfig("BookInfo", "xRes");
-	_screenHeight = getIntFromConfig("BookInfo", "yRes");
-	// nColors is here too, but it's always 256 anyway...
-
-	// this is 1 in The New Kid on the Block, changes the hardcoded UI
-	// v2 games changed the flag name to fPoetry
-	if (getGameType() == GType_LIVINGBOOKSV1)
-		_poetryMode = (getIntFromConfig("BookInfo", "poetry") == 1);
-	else
-		_poetryMode = (getIntFromConfig("BookInfo", "fPoetry") == 1);
-
-	// The later Living Books games add some more options:
-	//     - fNeedPalette                (always true?)
-	//     - fUse254ColorPalette         (always true?)
-	//     - nKBRequired                 (4096, RAM requirement?)
-	//     - fDebugWindow                (always 0?)
-
-	if (_bookInfoFile.hasSection("Globals")) {
-		const Common::INIFile::SectionKeyList globals = _bookInfoFile.getKeys("Globals");
-		for (Common::INIFile::SectionKeyList::const_iterator i = globals.begin(); i != globals.end(); i++) {
-			Common::String command = Common::String::format("%s = %s", i->key.c_str(), i->value.c_str());
-			LBCode tempCode(this, 0);
-			uint offset = tempCode.parseCode(command);
-			tempCode.runCode(nullptr, offset);
-		}
-	}
-}
-
-Common::String MohawkEngine_LivingBooks::stringForMode(LBMode mode) {
-	Common::String language = getStringFromConfig("Languages", Common::String::format("Language%d", _curLanguage));
-
-	switch (mode) {
-	case kLBIntroMode:
-		return "Intro";
-	case kLBControlMode:
-		return "Control";
-	case kLBCreditsMode:
-		return "Credits";
-	case kLBPreviewMode:
-		return "Preview";
-	case kLBReadMode:
-		return language + ".Read";
-	case kLBPlayMode:
-		return language + ".Play";
-	default:
-		error("unknown game mode %d", (int)mode);
 	}
 }
 
@@ -798,167 +683,9 @@ void MohawkEngine_LivingBooks::lockSound(LBItem *owner, bool lock) {
 	}
 }
 
-// Only 1 VSRN resource per page
-uint16 LBPage::getResourceVersion() {
-	Common::SeekableReadStream *versionStream = _vm->getResource(ID_VRSN, _baseId);
-
-	// FIXME: some V2 games have very strange version entries
-	if (versionStream->size() != 2)
-		debug(1, "Version Record size mismatch");
-
-	uint16 version = versionStream->readUint16BE();
-
-	delete versionStream;
-	return version;
-}
-
-void LBPage::loadBITL(uint16 resourceId) {
-	Common::SeekableReadStreamEndian *bitlStream = _vm->wrapStreamEndian(ID_BITL, resourceId);
-
-	while (true) {
-		Common::Rect rect = _vm->readRect(bitlStream);
-		uint16 type = bitlStream->readUint16();
-
-		LBItem *res;
-		switch (type) {
-		case kLBPictureItem:
-			res = new LBPictureItem(_vm, this, rect);
-			break;
-		case kLBAnimationItem:
-			res = new LBAnimationItem(_vm, this, rect);
-			break;
-		case kLBPaletteItem:
-			res = new LBPaletteItem(_vm, this, rect);
-			break;
-		case kLBGroupItem:
-			res = new LBGroupItem(_vm, this, rect);
-			break;
-		case kLBSoundItem:
-			res = new LBSoundItem(_vm, this, rect);
-			break;
-		case kLBLiveTextItem:
-			res = new LBLiveTextItem(_vm, this, rect);
-			break;
-		case kLBMovieItem:
-			res = new LBMovieItem(_vm, this, rect);
-			break;
-		case kLBMiniGameItem:
-			res = new LBMiniGameItem(_vm, this, rect);
-			break;
-		case kLBProxyItem:
-			res = new LBProxyItem(_vm, this, rect);
-			break;
-		default:
-			warning("Unknown item type %04x", type);
-			// fall through
-		case 3: // often used for buttons
-			res = new LBItem(_vm, this, rect);
-			break;
-		}
-
-		res->readFrom(bitlStream);
-		_items.push_back(res);
-
-		if (bitlStream->size() == bitlStream->pos())
-			break;
-	}
-
-	delete bitlStream;
-}
-
 Common::SeekableReadStreamEndian *MohawkEngine_LivingBooks::wrapStreamEndian(uint32 tag, uint16 id) {
 	Common::SeekableReadStream *dataStream = getResource(tag, id);
 	return new Common::SeekableReadStreamEndianWrapper(dataStream, isBigEndian(), DisposeAfterUse::YES);
-}
-
-Common::String MohawkEngine_LivingBooks::getStringFromConfig(const Common::String &section, const Common::String &key) {
-	Common::String x, leftover;
-	_bookInfoFile.getKey(key, section, x);
-	Common::String tmp = removeQuotesFromString(x, leftover);
-	if (!leftover.empty())
-		warning("while parsing config key '%s' from section '%s', string '%s' was left after '%s'",
-			key.c_str(), section.c_str(), leftover.c_str(), tmp.c_str());
-	return tmp;
-}
-
-Common::String MohawkEngine_LivingBooks::getStringFromConfig(const Common::String &section, const Common::String &key, Common::String &leftover) {
-	Common::String x;
-	_bookInfoFile.getKey(key, section, x);
-	return removeQuotesFromString(x, leftover);
-}
-
-int MohawkEngine_LivingBooks::getIntFromConfig(const Common::String &section, const Common::String &key) {
-	return atoi(getStringFromConfig(section, key).c_str());
-}
-
-Common::String MohawkEngine_LivingBooks::getFileNameFromConfig(const Common::String &section, const Common::String &key, Common::String &leftover) {
-	Common::String string = getStringFromConfig(section, key, leftover);
-
-	if (string.hasPrefix("//")) {
-		// skip "//CD-ROM Title/" prefixes which we don't care about
-		uint i = 3;
-		while (i < string.size() && string[i - 1] != '/')
-			i++;
-
-		// Already uses slashes, no need to convert further
-		return string.c_str() + i;
-	}
-
-	return (getPlatform() == Common::kPlatformMacintosh) ? convertMacFileName(string) : convertWinFileName(string);
-}
-
-Common::String MohawkEngine_LivingBooks::removeQuotesFromString(const Common::String &string, Common::String &leftover) {
-	if (string.empty())
-		return string;
-
-	char quoteChar = string[0];
-	if (quoteChar != '\"' && quoteChar != '\'')
-		return string;
-
-	Common::String tmp;
-	bool inLeftover = false;
-	for (uint32 i = 1; i < string.size(); i++) {
-		if (inLeftover)
-			leftover += string[i];
-		else if (string[i] == quoteChar)
-			inLeftover = true;
-		else
-			tmp += string[i];
-	}
-
-	return tmp;
-}
-
-Common::String MohawkEngine_LivingBooks::convertMacFileName(const Common::String &string) {
-	Common::String filename;
-
-	for (uint32 i = 0; i < string.size(); i++) {
-		if (i == 0 && string[i] == ':') // First character should be ignored (another colon)
-			continue;
-		if (string[i] == ':') // Directory separator
-			filename += '/';
-		else if (string[i] == '/') // Literal slash
-			filename += ':'; // Replace by colon, as used by macOS for slash
-		else
-			filename += string[i];
-	}
-
-	return filename;
-}
-
-Common::String MohawkEngine_LivingBooks::convertWinFileName(const Common::String &string) {
-	Common::String filename;
-
-	for (uint32 i = 0; i < string.size(); i++) {
-		if (i == 0 && (string[i] == '/' || string[i] == '\\')) // ignore slashes at start
-			continue;
-		if (string[i] == '\\')
-			filename += '/';
-		else
-			filename += string[i];
-	}
-
-	return filename;
 }
 
 Archive *MohawkEngine_LivingBooks::createArchive() const {
@@ -1030,285 +757,6 @@ void MohawkEngine_LivingBooks::nextPage() {
 		return;
 
 	error("Could not find page after %d.%d for mode %d", _curPage, _curSubPage, (int)_curMode);
-}
-
-void MohawkEngine_LivingBooks::handleUIMenuClick(uint controlId) {
-	LBItem *item;
-
-	switch (controlId) {
-	case 1:
-		if (getFeatures() & GF_LB_10) {
-			if (!tryLoadPageStart(kLBControlMode, 2))
-				error("couldn't load options page");
-		} else {
-			if (!tryLoadPageStart(kLBControlMode, 3))
-				error("couldn't load options page");
-		}
-		break;
-
-	case 2:
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(11);
-		if (item)
-			item->destroySelf();
-		item = getItemById(199 + _curLanguage);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false, true);
-		}
-		break;
-
-	case 3:
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(11);
-		if (item)
-			item->destroySelf();
-		item = getItemById(12);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false, true);
-		}
-		break;
-
-	case 4:
-		if (getFeatures() & GF_LB_10) {
-			if (!tryLoadPageStart(kLBControlMode, 3))
-				error("couldn't load quit page");
-		} else {
-			if (!tryLoadPageStart(kLBControlMode, 2))
-				error("couldn't load quit page");
-		}
-		break;
-
-	case 10:
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(11);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false);
-		}
-		break;
-
-	case 11:
-		item = getItemById(11);
-		if (item)
-			item->togglePlaying(false, true);
-		break;
-
-	case 12:
-		// start game, in play mode
-		if (!tryLoadPageStart(kLBPlayMode, 1))
-			error("couldn't start play mode");
-		break;
-
-	default:
-		if (controlId >= 100 && controlId < 100 + (uint)_numLanguages) {
-			uint newLanguage = controlId - 99;
-			if (newLanguage == _curLanguage)
-				break;
-			item = getItemById(99 + _curLanguage);
-			if (item)
-				item->seek(1);
-			_curLanguage = newLanguage;
-		} else if (controlId >= 200 && controlId < 200 + (uint)_numLanguages) {
-			// start game, in read mode
-			if (!tryLoadPageStart(kLBReadMode, 1))
-				error("couldn't start read mode");
-		}
-		break;
-	}
-}
-
-void MohawkEngine_LivingBooks::handleUIPoetryMenuClick(uint controlId) {
-	LBItem *item;
-
-	// the menu UI in New Kid on the Block is a hybrid of the normal menu
-	// and the normal options screen
-
-	// TODO: this is mostly untested
-
-	switch (controlId) {
-	case 2:
-	case 3:
-		handleUIOptionsClick(controlId);
-		break;
-
-	case 4:
-		handleUIMenuClick(controlId);
-		break;
-
-	case 6:
-		handleUIMenuClick(2);
-		break;
-
-	case 7:
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(11);
-		if (item)
-			item->destroySelf();
-		item = getItemById(12);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false, true);
-		}
-		break;
-
-	case 0xA:
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(11);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false);
-		}
-		break;
-
-	case 0xB:
-		item = getItemById(11);
-		if (item)
-			item->togglePlaying(false, true);
-		break;
-
-	case 0xC:
-		if (!tryLoadPageStart(kLBPlayMode, _curSelectedPage))
-			error("failed to load page %d", _curSelectedPage);
-		break;
-
-	default:
-		if (controlId < 100) {
-			handleUIMenuClick(controlId);
-		} else {
-			if (!tryLoadPageStart(kLBReadMode, _curSelectedPage))
-				error("failed to load page %d", _curSelectedPage);
-		}
-	}
-}
-
-void MohawkEngine_LivingBooks::handleUIQuitClick(uint controlId) {
-	LBItem *item;
-
-	switch (controlId) {
-	case 1:
-	case 2:
-		// button clicked, run animation
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(11);
-		if (item)
-			item->destroySelf();
-		item = getItemById((controlId == 1) ? 12 : 13);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false);
-		}
-		break;
-
-	case 10:
-	case 11:
-		item = getItemById(11);
-		if (item)
-			item->togglePlaying(false, true);
-		break;
-
-	case 12:
-		// 'yes', I want to quit
-		quitGame();
-		break;
-
-	case 13:
-		// 'no', go back to menu
-		if (!tryLoadPageStart(kLBControlMode, 1))
-			error("couldn't return to menu");
-		break;
-
-	default:
-		break;
-	}
-}
-
-void MohawkEngine_LivingBooks::handleUIOptionsClick(uint controlId) {
-	LBItem *item;
-
-	switch (controlId) {
-	case 1:
-		item = getItemById(10);
-		if (item)
-			item->destroySelf();
-		item = getItemById(202);
-		if (item) {
-			item->setVisible(true);
-			item->togglePlaying(false, true);
-		}
-		break;
-
-	case 2:
-		// back
-		item = getItemById(2);
-		if (item)
-			item->seek(1);
-		if (_curSelectedPage == 1) {
-			_curSelectedPage = _numPages;
-		} else {
-			_curSelectedPage--;
-		}
-		for (uint i = 0; i < _numPages; i++) {
-			item = getItemById(1000 + i);
-			if (item)
-				item->setVisible(_curSelectedPage == i + 1);
-			item = getItemById(1100 + i);
-			if (item)
-				item->setVisible(_curSelectedPage == i + 1);
-		}
-		break;
-
-	case 3:
-		// forward
-		item = getItemById(3);
-		if (item)
-			item->seek(1);
-		if (_curSelectedPage == _numPages) {
-			_curSelectedPage = 1;
-		} else {
-			_curSelectedPage++;
-		}
-		for (uint i = 0; i < _numPages; i++) {
-			item = getItemById(1000 + i);
-			if (item)
-				item->setVisible(_curSelectedPage == i + 1);
-			item = getItemById(1100 + i);
-			if (item)
-				item->setVisible(_curSelectedPage == i + 1);
-		}
-		break;
-
-	case 4:
-		if (!tryLoadPageStart(kLBCreditsMode, 1))
-			error("failed to start credits");
-		break;
-
-	case 5:
-		if (!tryLoadPageStart(kLBPreviewMode, 1))
-			error("failed to start preview");
-		break;
-
-	case 202:
-		if (!tryLoadPageStart(kLBPlayMode, _curSelectedPage))
-			error("failed to load page %d", _curSelectedPage);
-		break;
-
-	default:
-		break;
-	}
 }
 
 void MohawkEngine_LivingBooks::handleNotify(NotifyEvent &event) {
@@ -1467,567 +915,6 @@ void MohawkEngine_LivingBooks::handleNotify(NotifyEvent &event) {
 	default:
 		error("Unknown notification %d (param 0x%04x)", event.type, event.param);
 	}
-}
-
-LBAnimationNode::LBAnimationNode(MohawkEngine_LivingBooks *vm, LBAnimation *parent, uint16 scriptResourceId) : _vm(vm), _parent(parent) {
-	loadScript(scriptResourceId);
-}
-
-LBAnimationNode::~LBAnimationNode() {
-	for (uint32 i = 0; i < _scriptEntries.size(); i++)
-		delete[] _scriptEntries[i].data;
-}
-
-void LBAnimationNode::loadScript(uint16 resourceId) {
-	Common::SeekableReadStreamEndian *scriptStream = _vm->wrapStreamEndian(ID_SCRP, resourceId);
-
-	reset();
-
-	while (byte opcodeId = scriptStream->readByte()) {
-		byte size = scriptStream->readByte();
-
-		LBAnimScriptEntry entry;
-		entry.opcode = opcodeId;
-		entry.size = size;
-
-		if (!size) {
-			entry.data = nullptr;
-		} else {
-			entry.data = new byte[entry.size];
-			scriptStream->read(entry.data, entry.size);
-		}
-
-		_scriptEntries.push_back(entry);
-	}
-
-	byte size = scriptStream->readByte();
-	if (size != 0 || scriptStream->pos() != scriptStream->size())
-		error("Failed to read script correctly");
-
-	delete scriptStream;
-}
-
-void LBAnimationNode::draw(const Common::Rect &_bounds) {
-	if (!_currentCel)
-		return;
-
-	// this is also checked in SetCel, below
-	if (_currentCel > _parent->getNumResources())
-		error("Animation cel %d was too high, this shouldn't happen!", _currentCel);
-
-	int16 xOffset = _xPos + _bounds.left;
-	int16 yOffset = _yPos + _bounds.top;
-
-	uint16 resourceId = _parent->getResource(_currentCel - 1);
-
-	if (!_vm->isPreMohawk()) {
-		Common::Point offset = _parent->getOffset(_currentCel - 1);
-		xOffset -= offset.x;
-		yOffset -= offset.y;
-	}
-
-	_vm->_gfx->copyOffsetAnimImageToScreen(resourceId, xOffset, yOffset);
-}
-
-void LBAnimationNode::reset() {
-	// TODO: this causes stupid flickering
-	//if (_currentCel)
-	//	_vm->_needsRedraw = true;
-
-	_currentCel = 0;
-	_currentEntry = 0;
-	_delay = 0;
-
-	_xPos = 0;
-	_yPos = 0;
-}
-
-NodeState LBAnimationNode::update(bool seeking) {
-	if (_currentEntry == _scriptEntries.size())
-		return kLBNodeDone;
-
-	if (_delay > 0 && --_delay)
-		return kLBNodeRunning;
-
-	while (_currentEntry < _scriptEntries.size()) {
-		LBAnimScriptEntry &entry = _scriptEntries[_currentEntry];
-		_currentEntry++;
-		debug(5, "Running script entry %d of %d", _currentEntry, _scriptEntries.size());
-
-		switch (entry.opcode) {
-		case kLBAnimOpPlaySound:
-		case kLBAnimOpWaitForSound:
-		case kLBAnimOpReleaseSound:
-		case kLBAnimOpResetSound:
-			{
-			uint16 soundResourceId = READ_BE_UINT16(entry.data);
-
-			if (!soundResourceId) {
-				error("Unhandled named wave file, tell clone2727 where you found this");
-				break;
-			}
-
-			Common::String cue;
-			uint pos = 2;
-			while (pos < entry.size) {
-				char in = entry.data[pos];
-				if (!in)
-					break;
-				pos++;
-				cue += in;
-			}
-			if (pos == entry.size)
-				error("Cue in sound kLBAnimOp wasn't null-terminated");
-
-			switch (entry.opcode) {
-			case kLBAnimOpPlaySound:
-				if (seeking)
-					break;
-				debug(4, "a: PlaySound(%0d)", soundResourceId);
-				_parent->playSound(soundResourceId);
-				break;
-			case kLBAnimOpWaitForSound:
-				if (seeking)
-					break;
-				debug(4, "b: WaitForSound(%0d)", soundResourceId);
-				if (!_parent->soundPlaying(soundResourceId, cue))
-					break;
-				_currentEntry--;
-				return kLBNodeWaiting;
-			case kLBAnimOpReleaseSound:
-				debug(4, "c: ReleaseSound(%0d)", soundResourceId);
-				// TODO
-				_vm->_sound->stopSound(soundResourceId);
-				break;
-			case kLBAnimOpResetSound:
-				debug(4, "d: ResetSound(%0d)", soundResourceId);
-				// TODO
-				_vm->_sound->stopSound(soundResourceId);
-				break;
-			default:
-				break;
-			}
-			}
-			break;
-
-		case kLBAnimOpSetTempo:
-		case kLBAnimOpSetTempoDiv:
-			{
-			assert(entry.size == 2);
-			uint16 tempo = (int16)READ_BE_UINT16(entry.data);
-
-			// TODO: LB 3 uses fixed-point here.
-			if (entry.opcode == kLBAnimOpSetTempo) {
-				debug(4, "3: SetTempo(%d)", tempo);
-				// TODO: LB 3 uses (tempo * 1000) / 60, while
-				// the original divides the system time by 16.
-				_parent->setTempo(tempo * 16);
-			} else {
-				// LB 3.0+ only.
-				debug(4, "E: SetTempoDiv(%d)", tempo);
-				_parent->setTempo(1000 / tempo);
-			}
-
-			}
-			break;
-
-		case kLBAnimOpWait:
-			assert(entry.size == 0);
-			debug(5, "6: Wait()");
-			return kLBNodeRunning;
-
-		case kLBAnimOpMoveTo:
-			{
-			assert(entry.size == 4);
-			int16 x = (int16)READ_BE_UINT16(entry.data);
-			int16 y = (int16)READ_BE_UINT16(entry.data + 2);
-			debug(4, "5: MoveTo(%d, %d)", x, y);
-
-			_xPos = x;
-			_yPos = y;
-			_vm->_needsRedraw = true;
-			}
-			break;
-
-		case kLBAnimOpDrawMode:
-			{
-			assert(entry.size == 2);
-			uint16 mode = (int16)READ_BE_UINT16(entry.data);
-			debug(4, "9: DrawMode(%d)", mode);
-
-			// TODO
-			}
-			break;
-
-		case kLBAnimOpSetCel:
-			{
-			assert(entry.size == 2);
-			uint16 cel = (int16)READ_BE_UINT16(entry.data);
-			debug(4, "7: SetCel(%d)", cel);
-
-			_currentCel = cel;
-			if (_currentCel > _parent->getNumResources())
-				error("SetCel set current cel to %d, but we only have %d cels", _currentCel, _parent->getNumResources());
-			_vm->_needsRedraw = true;
-			}
-			break;
-
-		case kLBAnimOpNotify:
-			{
-			assert(entry.size == 2);
-			uint16 data = (int16)READ_BE_UINT16(entry.data);
-
-			if (seeking)
-				break;
-
-			debug(4, "2: Notify(%d)", data);
-			_vm->notifyAll(data, _parent->getParentId());
-			}
-			break;
-
-		case kLBAnimOpSleepUntil:
-			{
-			assert(entry.size == 4);
-			uint32 frame = READ_BE_UINT32(entry.data);
-			debug(4, "8: SleepUntil(%d)", frame);
-
-			if (frame > _parent->getCurrentFrame()) {
-				// *not* kLBNodeWaiting
-				_currentEntry--;
-				return kLBNodeRunning;
-			}
-			}
-			break;
-
-		case kLBAnimOpDelay:
-			{
-			assert(entry.size == 4);
-			uint32 delay = READ_BE_UINT32(entry.data);
-			debug(4, "f: Delay(%d)", delay);
-			_delay = delay;
-			return kLBNodeRunning;
-			}
-			break;
-
-		default:
-			error("Unknown opcode id %02x (size %d)", entry.opcode, entry.size);
-			break;
-		}
-	}
-
-	return kLBNodeRunning;
-}
-
-bool LBAnimationNode::transparentAt(int x, int y) {
-	if (!_currentCel)
-		return true;
-
-	uint16 resourceId = _parent->getResource(_currentCel - 1);
-
-	if (!_vm->isPreMohawk()) {
-		Common::Point offset = _parent->getOffset(_currentCel - 1);
-		x += offset.x;
-		y += offset.y;
-	}
-
-	// TODO: only check pixels if necessary
-	return _vm->_gfx->imageIsTransparentAt(resourceId, true, x - _xPos, y - _yPos);
-}
-
-LBAnimation::LBAnimation(MohawkEngine_LivingBooks *vm, LBAnimationItem *parent, uint16 resourceId) : _vm(vm), _parent(parent) {
-	Common::SeekableReadStreamEndian *aniStream = _vm->wrapStreamEndian(ID_ANI, resourceId);
-
-	// ANI records in the Wanderful sampler are 32 bytes, extra bytes are just NULs
-	if (aniStream->size() != 30 && aniStream->size() != 32)
-		warning("ANI Record size mismatch");
-
-	uint16 version = aniStream->readUint16();
-	if (version != 1)
-		warning("ANI version not 1");
-
-	_bounds = _vm->readRect(aniStream);
-	_clip = _vm->readRect(aniStream);
-	// TODO: what is colorId for?
-	uint32 colorId = aniStream->readUint32();
-	uint32 sprResourceId = aniStream->readUint32();
-	uint32 sprResourceOffset = aniStream->readUint32();
-
-	debug(5, "ANI bounds: (%d, %d), (%d, %d)", _bounds.left, _bounds.top, _bounds.right, _bounds.bottom);
-	debug(5, "ANI clip: (%d, %d), (%d, %d)", _clip.left, _clip.top, _clip.right, _clip.bottom);
-	debug(5, "ANI color id: %d", colorId);
-	debug(5, "ANI SPRResourceId: %d, offset %d", sprResourceId, sprResourceOffset);
-	if (aniStream->size() == 32) {
-		debug(5, "ANI extra bytes: (%d)", aniStream->readUint16());
-	}
-
-	if (aniStream->pos() != aniStream->size())
-		error("Still %d bytes at the end of anim stream", (int)(aniStream->size() - aniStream->pos()));
-
-	delete aniStream;
-
-	if (sprResourceOffset)
-		error("Cannot handle non-zero ANI offset yet");
-
-	Common::SeekableReadStreamEndian *sprStream = _vm->wrapStreamEndian(ID_SPR, sprResourceId);
-
-	uint16 numBackNodes = sprStream->readUint16();
-	uint16 numFrontNodes = sprStream->readUint16();
-	uint32 shapeResourceID = sprStream->readUint32();
-	uint32 shapeResourceOffset = sprStream->readUint32();
-	uint32 scriptResourceID = sprStream->readUint32();
-	uint32 scriptResourceOffset = sprStream->readUint32();
-	uint32 scriptResourceLength = sprStream->readUint32();
-	debug(5, "SPR# stream: %d front, %d background", numFrontNodes, numBackNodes);
-	debug(5, "Shape ID %d (offset 0x%04x), script ID %d (offset 0x%04x, length %d)", shapeResourceID, shapeResourceOffset,
-		scriptResourceID, scriptResourceOffset, scriptResourceLength);
-
-	Common::Array<uint16> scriptIDs;
-	for (uint16 i = 0; i < numFrontNodes; i++) {
-		uint32 unknown1 = sprStream->readUint32();
-		uint32 unknown2 = sprStream->readUint32();
-		uint32 unknown3 = sprStream->readUint32();
-		uint16 scriptID = sprStream->readUint32();
-		uint32 unknown4 = sprStream->readUint32();
-		uint32 unknown5 = sprStream->readUint32();
-		scriptIDs.push_back(scriptID);
-		debug(6, "Front node %d: script ID %d", i, scriptID);
-		if (unknown1 != 0 || unknown2 != 0 || unknown3 != 0 || unknown4 != 0 || unknown5 != 0)
-			error("Anim node %d had non-zero unknowns %08x, %08x, %08x, %08x, %08x",
-				i, unknown1, unknown2, unknown3, unknown4, unknown5);
-	}
-
-	if (numBackNodes)
-		error("Ignoring %d back nodes", numBackNodes);
-
-	if (sprStream->pos() != sprStream->size())
-		error("Still %d bytes at the end of sprite stream", (int)(sprStream->size() - sprStream->pos()));
-
-	delete sprStream;
-
-	loadShape(shapeResourceID);
-
-	_nodes.push_back(new LBAnimationNode(_vm, this, scriptResourceID));
-	for (uint16 i = 0; i < scriptIDs.size(); i++)
-		_nodes.push_back(new LBAnimationNode(_vm, this, scriptIDs[i]));
-
-	_currentFrame = 0;
-	_currentSound = 0xffff;
-	_running = false;
-	_tempo = 1;
-}
-
-LBAnimation::~LBAnimation() {
-	for (uint32 i = 0; i < _nodes.size(); i++)
-		delete _nodes[i];
-	if (_currentSound != 0xffff)
-		_vm->_sound->stopSound(_currentSound);
-}
-
-void LBAnimation::loadShape(uint16 resourceId) {
-	if (resourceId == 0)
-		return;
-
-	Common::SeekableReadStreamEndian *shapeStream = _vm->wrapStreamEndian(ID_SHP, resourceId);
-
-	if (_vm->isPreMohawk()) {
-		if (shapeStream->size() < 6)
-			error("V1 SHP Record size too short (%d)", (int)shapeStream->size());
-
-		uint16 u0 = shapeStream->readUint16();
-		if (u0 != 3)
-			error("V1 SHP Record u0 is %04x, not 3", u0);
-
-		uint16 u1 = shapeStream->readUint16();
-		if (u1 != 0)
-			error("V1 SHP Record u1 is %04x, not 0", u1);
-
-		uint16 idCount = shapeStream->readUint16();
-		debug(8, "V1 SHP: idCount: %d", idCount);
-
-		if (shapeStream->size() != (idCount * 2) + 6)
-			error("V1 SHP Record size mismatch (%d)", (int)shapeStream->size());
-
-		for (uint16 i = 0; i < idCount; i++) {
-			_shapeResources.push_back(shapeStream->readUint16());
-			debug(8, "V1 SHP: BMAP Resource Id %d: %d", i, _shapeResources[i]);
-		}
-	} else {
-		uint16 idCount = shapeStream->readUint16();
-		debug(8, "SHP: idCount: %d", idCount);
-
-		if (shapeStream->size() != (idCount * 6) + 2)
-			error("SHP Record size mismatch (%d)", (int)shapeStream->size());
-
-		for (uint16 i = 0; i < idCount; i++) {
-			_shapeResources.push_back(shapeStream->readUint16());
-			int16 x = shapeStream->readSint16();
-			int16 y = shapeStream->readSint16();
-			_shapeOffsets.push_back(Common::Point(x, y));
-			debug(8, "SHP: tBMP Resource Id %d: %d, at (%d, %d)", i, _shapeResources[i], x, y);
-		}
-	}
-
-	for (uint16 i = 0; i < _shapeResources.size(); i++)
-		_vm->_gfx->preloadImage(_shapeResources[i]);
-
-	delete shapeStream;
-}
-
-void LBAnimation::draw() {
-	for (uint32 i = 0; i < _nodes.size(); i++)
-		_nodes[i]->draw(_bounds);
-}
-
-bool LBAnimation::update() {
-	if (!_running)
-		return false;
-
-	if (_vm->_system->getMillis() <= _lastTime + (uint32)_tempo)
-		return false;
-
-	// the second check is to try 'catching up' with lagged animations, might be crazy
-	if (_lastTime == 0 || (_vm->_system->getMillis()) > _lastTime + (uint32)(_tempo * 2))
-		_lastTime = _vm->_system->getMillis();
-	else
-		_lastTime += _tempo;
-
-	if (_currentSound != 0xffff && !_vm->_sound->isPlaying(_currentSound)) {
-		_currentSound = 0xffff;
-	}
-
-	NodeState state = kLBNodeDone;
-	for (uint32 i = 0; i < _nodes.size(); i++) {
-		NodeState s = _nodes[i]->update();
-		if (s == kLBNodeWaiting) {
-			state = kLBNodeWaiting;
-			if (i != 0)
-				warning("non-primary node was waiting");
-			break;
-		}
-		if (s == kLBNodeRunning)
-			state = kLBNodeRunning;
-	}
-
-	if (state == kLBNodeRunning) {
-		_currentFrame++;
-	} else if (state == kLBNodeDone) {
-		if (_currentSound == 0xffff) {
-			_running = false;
-			return true;
-		}
-	}
-
-	return false;
-}
-
-void LBAnimation::start() {
-	_lastTime = 0;
-	_running = true;
-}
-
-void LBAnimation::seek(uint16 pos) {
-	_lastTime = 0;
-	_currentFrame = 0;
-
-	if (_currentSound != 0xffff) {
-		_vm->_sound->stopSound(_currentSound);
-		_currentSound = 0xffff;
-	}
-
-	for (uint32 i = 0; i < _nodes.size(); i++)
-		_nodes[i]->reset();
-
-	for (uint16 n = 0; n < pos; n++) {
-		bool ranSomething = false;
-		// nodes don't wait while seeking
-		for (uint32 i = 0; i < _nodes.size(); i++)
-			ranSomething |= (_nodes[i]->update(true) != kLBNodeDone);
-
-		_currentFrame++;
-
-		if (!ranSomething) {
-			_running = false;
-			break;
-		}
-	}
-}
-
-void LBAnimation::seekToTime(uint32 time) {
-	_lastTime = 0;
-	_currentFrame = 0;
-
-	if (_currentSound != 0xffff) {
-		_vm->_sound->stopSound(_currentSound);
-		_currentSound = 0xffff;
-	}
-
-	for (uint32 i = 0; i < _nodes.size(); i++)
-		_nodes[i]->reset();
-
-	uint32 elapsed = 0;
-	while (elapsed <= time) {
-		bool ranSomething = false;
-		// nodes don't wait while seeking
-		for (uint32 i = 0; i < _nodes.size(); i++)
-			ranSomething |= (_nodes[i]->update(true) != kLBNodeDone);
-
-		elapsed += _tempo;
-		_currentFrame++;
-
-		if (!ranSomething) {
-			_running = false;
-			break;
-		}
-	}
-}
-
-void LBAnimation::stop() {
-	_running = false;
-	if (_currentSound != 0xffff) {
-		_vm->_sound->stopSound(_currentSound);
-		_currentSound = 0xffff;
-	}
-}
-
-void LBAnimation::playSound(uint16 resourceId) {
-	_currentSound = resourceId;
-	_vm->_sound->playSound(_currentSound, Audio::Mixer::kMaxChannelVolume, false, &_cueList);
-}
-
-bool LBAnimation::soundPlaying(uint16 resourceId, const Common::String &cue) {
-	if (_currentSound != resourceId)
-		return false;
-	if (!_vm->_sound->isPlaying(_currentSound))
-		return false;
-
-	if (cue.empty())
-		return true;
-
-	uint samples = _vm->_sound->getNumSamplesPlayed(_currentSound);
-	for (uint i = 0; i < _cueList.pointCount; i++) {
-		if (_cueList.points[i].sampleFrame > samples)
-			break;
-		if (_cueList.points[i].name == cue)
-			return false;
-	}
-
-	return true;
-}
-
-bool LBAnimation::transparentAt(int x, int y) {
-	for (uint32 i = 0; i < _nodes.size(); i++)
-		if (!_nodes[i]->transparentAt(x - _bounds.left, y - _bounds.top))
-			return false;
-
-	return true;
-}
-
-void LBAnimation::setTempo(uint16 tempo) {
-	_tempo = tempo;
-}
-
-uint16 LBAnimation::getParentId() {
-	return _parent->getId();
 }
 
 LBScriptEntry::LBScriptEntry() {
@@ -3144,859 +2031,6 @@ bool LBItem::checkCondition(const Common::String &condition) {
 	LBValue result = tempCode.runCode(this, offset);
 
 	return result.toInt();
-}
-
-LBSoundItem::LBSoundItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBSoundItem");
-	_running = false;
-}
-
-LBSoundItem::~LBSoundItem() {
-	if (_running)
-		_vm->_sound->stopSound(_resourceId);
-}
-
-void LBSoundItem::update() {
-	if (_running && !_vm->_sound->isPlaying(_resourceId)) {
-		_running = false;
-		done(true);
-	}
-
-	LBItem::update();
-}
-
-bool LBSoundItem::togglePlaying(bool playing, bool restart) {
-	if (!playing)
-		return LBItem::togglePlaying(playing, restart);
-
-	if (_running) {
-		_running = false;
-		_vm->_sound->stopSound(_resourceId);
-	}
-
-	if (!_loaded || !_enabled || !_globalEnabled)
-		return false;
-
-	_running = true;
-	debug(4, "sound %d play for item %d (%s)", _resourceId, _itemId, _desc.c_str());
-	_vm->playSound(this, _resourceId);
-	return true;
-}
-
-void LBSoundItem::stop() {
-	if (_running) {
-		_running = false;
-		_vm->_sound->stopSound(_resourceId);
-	}
-
-	LBItem::stop();
-}
-
-LBItem *LBSoundItem::createClone() {
-	return new LBSoundItem(_vm, _page, _rect);
-}
-
-LBGroupItem::LBGroupItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBGroupItem");
-	_starting = false;
-}
-
-void LBGroupItem::readData(uint16 type, uint16 size, Common::MemoryReadStreamEndian *stream) {
-	switch (type) {
-	case kLBGroupData:
-		{
-		_groupEntries.clear();
-		uint16 count = stream->readUint16();
-		debug(3, "Group data: %d entries", count);
-
-		if (size != 2 + count * 4)
-			error("kLBGroupData was wrong size (%d, for %d entries)", size, count);
-
-		for (uint i = 0; i < count; i++) {
-			GroupEntry entry;
-			// TODO: is type important for any game? at the moment, we ignore it
-			entry.entryType = stream->readUint16();
-			entry.entryId = stream->readUint16();
-			// HACK: The Living Books v3 sampler includes the ID for the group as
-			// one of the entries in the Green Eggs and Ham section, which leads
-			// to infinite recursion when the group is loaded.
-			if (entry.entryId != getId())
-				_groupEntries.push_back(entry);
-			debug(3, "group entry: id %d, type %d", entry.entryId, entry.entryType);
-		}
-		}
-		break;
-
-	default:
-		LBItem::readData(type, size, stream);
-	}
-}
-
-void LBGroupItem::destroySelf() {
-	LBItem::destroySelf();
-
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->destroySelf();
-	}
-}
-
-void LBGroupItem::setEnabled(bool enabled) {
-	if (_starting) {
-		_starting = false;
-		LBItem::setEnabled(enabled);
-	} else {
-		for (uint i = 0; i < _groupEntries.size(); i++) {
-			LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-			if (item)
-				item->setEnabled(enabled);
-		}
-	}
-}
-
-void LBGroupItem::setGlobalEnabled(bool enabled) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->setGlobalEnabled(enabled);
-	}
-}
-
-bool LBGroupItem::contains(Common::Point point) {
-	return false;
-}
-
-bool LBGroupItem::togglePlaying(bool playing, bool restart) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->togglePlaying(playing, restart);
-	}
-
-	return false;
-}
-
-void LBGroupItem::seek(uint16 pos) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->seek(pos);
-	}
-}
-
-void LBGroupItem::setVisible(bool visible) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->setVisible(visible);
-	}
-}
-
-void LBGroupItem::setGlobalVisible(bool visible) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->setGlobalVisible(visible);
-	}
-}
-
-void LBGroupItem::startPhase(uint phase) {
-	_starting = true;
-	LBItem::startPhase(phase);
-	_starting = false;
-}
-
-void LBGroupItem::stop() {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->stop();
-	}
-}
-
-void LBGroupItem::load() {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->load();
-	}
-}
-
-void LBGroupItem::unload() {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->unload();
-	}
-}
-
-void LBGroupItem::moveBy(const Common::Point &pos) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->moveBy(pos);
-	}
-}
-
-void LBGroupItem::moveTo(const Common::Point &pos) {
-	for (uint i = 0; i < _groupEntries.size(); i++) {
-		LBItem *item = _vm->getItemById(_groupEntries[i].entryId);
-		if (item)
-			item->moveTo(pos);
-	}
-}
-
-LBItem *LBGroupItem::createClone() {
-	// TODO: needed?
-	error("LBGroupItem::createClone unimplemented");
-	return new LBGroupItem(_vm, _page, _rect);
-}
-
-LBPaletteItem::LBPaletteItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBPaletteItem");
-
-	_fadeInStart = 0;
-	_palette = nullptr;
-}
-
-LBPaletteItem::~LBPaletteItem() {
-	delete[] _palette;
-}
-
-void LBPaletteItem::readData(uint16 type, uint16 size, Common::MemoryReadStreamEndian *stream) {
-	switch (type) {
-	case kLBPaletteXData:
-		{
-		assert(size >= 8);
-		_fadeInPeriod = stream->readUint16();
-		_fadeInStep = stream->readUint16();
-		_drawStart = stream->readUint16();
-		_drawCount = stream->readUint16();
-		if (_drawStart + _drawCount > 256)
-			error("encountered palette trying to set more than 256 colors");
-		assert(size == 8 + _drawCount * 4);
-
-		// TODO: _drawCount is really more like _drawEnd, so once we're sure that
-		// there's really no use for the palette entries before _drawCount, we
-		// might want to just discard them here, at load time.
-		_palette = new byte[_drawCount * 3];
-		for (uint i = 0; i < _drawCount; i++) {
-			_palette[i*3 + 0] = stream->readByte();
-			_palette[i*3 + 1] = stream->readByte();
-			_palette[i*3 + 2] = stream->readByte();
-			stream->readByte();
-		}
-		}
-		break;
-
-	default:
-		LBItem::readData(type, size, stream);
-	}
-}
-
-bool LBPaletteItem::togglePlaying(bool playing, bool restart) {
-	// TODO: this likely isn't the right place
-
-	if (playing) {
-		_fadeInStart = _vm->_system->getMillis();
-		_fadeInCurrent = 0;
-
-		return true;
-	}
-
-	return LBItem::togglePlaying(playing, restart);
-}
-
-void LBPaletteItem::update() {
-	if (_fadeInStart) {
-		if (!_palette)
-			error("LBPaletteItem had no palette on startup");
-
-		uint32 elapsedTime = _vm->_system->getMillis() - _fadeInStart;
-		uint32 divTime = elapsedTime / _fadeInStep;
-
-		if (divTime > _fadeInPeriod)
-			divTime = _fadeInPeriod;
-
-		if (_fadeInCurrent != divTime) {
-			_fadeInCurrent = divTime;
-
-			// TODO: actual fading-in
-			if (_visible && _globalVisible) {
-				_vm->_system->getPaletteManager()->setPalette(_palette + _drawStart * 3, _drawStart, _drawCount - _drawStart);
-				_vm->_needsRedraw = true;
-			}
-		}
-
-		if (elapsedTime >= (uint32)_fadeInPeriod * (uint32)_fadeInStep) {
-			// TODO: correct?
-			_fadeInStart = 0;
-		}
-	}
-
-	LBItem::update();
-}
-
-LBItem *LBPaletteItem::createClone() {
-	error("can't clone LBPaletteItem");
-}
-
-LBLiveTextItem::LBLiveTextItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	_currentPhrase = 0xFFFF;
-	_currentWord = 0xFFFF;
-	debug(3, "new LBLiveTextItem");
-}
-
-void LBLiveTextItem::readData(uint16 type, uint16 size, Common::MemoryReadStreamEndian *stream) {
-	switch (type) {
-	case kLBLiveTextData:
-		{
-		stream->read(_backgroundColor, 4); // unused?
-		stream->read(_foregroundColor, 4);
-		stream->read(_highlightColor, 4);
-		_paletteIndex = stream->readUint16();
-		uint16 phraseCount = stream->readUint16();
-		uint16 wordCount = stream->readUint16();
-
-		debug(3, "LiveText has %d words in %d phrases, palette index 0x%04x", wordCount, phraseCount, _paletteIndex);
-		debug(3, "LiveText colors: background %02x%02x%02x%02x, foreground %02x%02x%02x%02x, highlight %02x%02x%02x%02x",
-			_backgroundColor[0], _backgroundColor[1], _backgroundColor[2], _backgroundColor[3],
-			_foregroundColor[0], _foregroundColor[1], _foregroundColor[2], _foregroundColor[3],
-			_highlightColor[0], _highlightColor[1], _highlightColor[2], _highlightColor[3]);
-
-		if (size != 18 + 14 * wordCount + 18 * phraseCount)
-			error("Bad Live Text data size (got %d, wanted %d words and %d phrases)", size, wordCount, phraseCount);
-
-		_words.clear();
-		for (uint i = 0; i < wordCount; i++) {
-			LiveTextWord word;
-			word.bounds = _vm->readRect(stream);
-			word.soundId = stream->readUint16();
-			word.itemType = stream->readUint16();
-			word.itemId = stream->readUint16();
-			debug(4, "Word: (%d, %d) to (%d, %d), sound %d, item %d (type %d)",
-				word.bounds.left, word.bounds.top, word.bounds.right, word.bounds.bottom, word.soundId, word.itemId, word.itemType);
-			_words.push_back(word);
-		}
-
-		_phrases.clear();
-		for (uint i = 0; i < phraseCount; i++) {
-			LiveTextPhrase phrase;
-			phrase.wordStart = stream->readUint16();
-			phrase.wordCount = stream->readUint16();
-			phrase.highlightStart = stream->readUint16();
-			phrase.startId = stream->readUint16();
-			phrase.highlightEnd = stream->readUint16();
-			phrase.endId = stream->readUint16();
-
-			// The original stored the values in uint32's so we need to swap here
-			if (_vm->isBigEndian()) {
-				SWAP(phrase.highlightStart, phrase.startId);
-				SWAP(phrase.highlightEnd, phrase.endId);
-			}
-
-			uint32 unknown1 = stream->readUint16();
-			uint16 unknown2 = stream->readUint32();
-
-			if (unknown1 != 0 || unknown2 != 0)
-				error("Unexpected unknowns %08x/%04x in LiveText word", unknown1, unknown2);
-
-			debug(4, "Phrase: start %d, count %d, start at %d (from %d), end at %d (from %d)",
-				phrase.wordStart, phrase.wordCount, phrase.highlightStart, phrase.startId, phrase.highlightEnd, phrase.endId);
-
-			_phrases.push_back(phrase);
-		}
-		}
-		break;
-
-	default:
-		LBItem::readData(type, size, stream);
-	}
-}
-
-bool LBLiveTextItem::contains(Common::Point point) {
-	if (!LBItem::contains(point))
-		return false;
-
-	point.x -= _rect.left;
-	point.y -= _rect.top;
-
-	for (uint i = 0; i < _words.size(); i++) {
-		if (_words[i].bounds.contains(point))
-			return true;
-	}
-
-	return false;
-}
-
-void LBLiveTextItem::paletteUpdate(uint16 word, bool on) {
-	_vm->_needsRedraw = true;
-
-	// Sometimes the last phrase goes out-of-bounds, the original engine
-	// only checks the words which are valid in the palette updating code.
-	if (word >= _words.size())
-		return;
-
-	if (_resourceId) {
-		// with a resource, we draw a bitmap in draw() rather than changing the palette
-		return;
-	}
-
-	if (on) {
-		_vm->_system->getPaletteManager()->setPalette(_highlightColor, _paletteIndex + word, 1);
-	} else {
-		_vm->_system->getPaletteManager()->setPalette(_foregroundColor, _paletteIndex + word, 1);
-	}
-}
-
-void LBLiveTextItem::update() {
-	if (_currentWord != 0xFFFF) {
-		uint16 soundId = _words[_currentWord].soundId;
-		if (soundId && !_vm->_sound->isPlaying(soundId)) {
-			paletteUpdate(_currentWord, false);
-
-			// TODO: check this in RE
-			LBItem *item = _vm->getItemById(_words[_currentWord].itemId);
-			if (item)
-				item->togglePlaying(false, true);
-
-			_currentWord = 0xFFFF;
-		}
-	}
-
-	LBItem::update();
-}
-
-void LBLiveTextItem::draw() {
-	// this is only necessary when we are drawing using a bitmap
-	if (!_resourceId)
-		return;
-
-	if (_currentWord != 0xFFFF) {
-		uint yPos = 0;
-		if (_currentWord > 0) {
-			for (uint i = 0; i < _currentWord; i++) {
-				yPos += (_words[i].bounds.bottom - _words[i].bounds.top);
-			}
-		}
-		drawWord(_currentWord, yPos);
-		return;
-	}
-
-	if (_currentPhrase == 0xFFFF)
-		return;
-
-	uint wordStart = _phrases[_currentPhrase].wordStart;
-	uint wordCount = _phrases[_currentPhrase].wordCount;
-	if (wordStart + wordCount > _words.size())
-		error("phrase %d was invalid (%d words, from %d, out of only %d total)",
-			_currentPhrase, wordCount, wordStart, _words.size());
-
-	uint yPos = 0;
-	for (uint i = 0; i < wordStart + wordCount; i++) {
-		if (i >= wordStart)
-			drawWord(i, yPos);
-		yPos += (_words[i].bounds.bottom - _words[i].bounds.top);
-	}
-}
-
-void LBLiveTextItem::drawWord(uint word, uint yPos) {
-	Common::Rect srcRect(0, yPos, _words[word].bounds.right - _words[word].bounds.left,
-		yPos + _words[word].bounds.bottom - _words[word].bounds.top);
-	Common::Rect dstRect = _words[word].bounds;
-	dstRect.translate(_rect.left, _rect.top);
-	_vm->_gfx->copyAnimImageSectionToScreen(_resourceId, srcRect, dstRect);
-}
-
-void LBLiveTextItem::handleMouseDown(Common::Point pos) {
-	if (!_loaded || !_enabled || !_globalEnabled || _playing)
-		return LBItem::handleMouseDown(pos);
-
-	pos.x -= _rect.left;
-	pos.y -= _rect.top;
-
-	for (uint i = 0; i < _words.size(); i++) {
-		if (_words[i].bounds.contains(pos)) {
-			if (_currentWord != 0xFFFF) {
-				paletteUpdate(_currentWord, false);
-				_currentWord = 0xFFFF;
-			}
-			uint16 soundId = _words[i].soundId;
-			if (!soundId) {
-				// TODO: can we be smarter here, using timing?
-				warning("ignoring click due to no soundId");
-				return;
-			}
-			_currentWord = i;
-			_vm->playSound(this, soundId);
-			paletteUpdate(_currentWord, true);
-			return;
-		}
-	}
-
-	return LBItem::handleMouseDown(pos);
-}
-
-bool LBLiveTextItem::togglePlaying(bool playing, bool restart) {
-	if (!playing)
-		return LBItem::togglePlaying(playing, restart);
-	if (!_loaded || !_enabled || !_globalEnabled)
-		return _playing;
-
-	// TODO: handle this properly
-	_vm->_sound->stopSound();
-
-	_currentWord = 0xFFFF;
-	_currentPhrase = 0xFFFF;
-
-	return true;
-}
-
-void LBLiveTextItem::stop() {
-	// TODO: stop sound, refresh palette
-
-	LBItem::stop();
-}
-
-void LBLiveTextItem::notify(uint16 data, uint16 from) {
-	if (!_loaded || !_enabled || !_globalEnabled || !_playing)
-		return LBItem::notify(data, from);
-
-	if (_currentWord != 0xFFFF) {
-		// TODO: handle this properly
-		_vm->_sound->stopSound();
-		paletteUpdate(_currentWord, false);
-		_currentWord = 0xFFFF;
-	}
-
-	for (uint i = 0; i < _phrases.size(); i++) {
-		if (_phrases[i].highlightStart == data && _phrases[i].startId == from) {
-			debug(2, "Enabling phrase %d", i);
-			for (uint j = 0; j < _phrases[i].wordCount; j++) {
-				paletteUpdate(_phrases[i].wordStart + j, true);
-			}
-			_currentPhrase = i;
-			// TODO: not sure this is the correct logic
-			if (i == _phrases.size() - 1) {
-				_currentPhrase = 0xFFFF;
-				done(true);
-			}
-		} else if (_phrases[i].highlightEnd == data && _phrases[i].endId == from) {
-			debug(2, "Disabling phrase %d", i);
-			for (uint j = 0; j < _phrases[i].wordCount; j++) {
-				paletteUpdate(_phrases[i].wordStart + j, false);
-			}
-			_currentPhrase = 0xFFFF;
-		}
-	}
-
-	LBItem::notify(data, from);
-}
-
-LBItem *LBLiveTextItem::createClone() {
-	error("can't clone LBLiveTextItem");
-}
-
-LBPictureItem::LBPictureItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBPictureItem");
-}
-
-void LBPictureItem::readData(uint16 type, uint16 size, Common::MemoryReadStreamEndian *stream) {
-	switch (type) {
-	case kLBSetDrawMode:
-		{
-		assert(size == 2);
-		// TODO: this probably sets whether points are always contained (0x10)
-		// or whether the bitmap contents are checked (00, or anything else?)
-		uint16 val = stream->readUint16();
-		debug(2, "LBPictureItem: kLBSetDrawMode: %04x", val);
-		}
-		break;
-
-	default:
-		LBItem::readData(type, size, stream);
-	}
-}
-
-bool LBPictureItem::contains(Common::Point point) {
-	if (!LBItem::contains(point))
-		return false;
-
-	if (!_doHitTest)
-		return true;
-
-	// TODO: only check pixels if necessary
-	return !_vm->_gfx->imageIsTransparentAt(_resourceId, false, point.x - _rect.left, point.y - _rect.top);
-}
-
-void LBPictureItem::init() {
-	_vm->_gfx->preloadImage(_resourceId);
-
-	LBItem::init();
-}
-
-void LBPictureItem::draw() {
-	if (!_loaded || !_visible || !_globalVisible)
-		return;
-
-	_vm->_gfx->copyAnimImageToScreen(_resourceId, _rect.left, _rect.top);
-}
-
-LBItem *LBPictureItem::createClone() {
-	return new LBPictureItem(_vm, _page, _rect);
-}
-
-LBAnimationItem::LBAnimationItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	_anim = nullptr;
-	_running = false;
-	debug(3, "new LBAnimationItem");
-}
-
-LBAnimationItem::~LBAnimationItem() {
-	delete _anim;
-}
-
-void LBAnimationItem::setEnabled(bool enabled) {
-	if (_running) {
-		if (enabled && _globalEnabled && !_loaded)
-			_anim->start();
-		else if (_loaded && !enabled && _enabled && _globalEnabled)
-			_anim->stop();
-	}
-
-	return LBItem::setEnabled(enabled);
-}
-
-bool LBAnimationItem::contains(Common::Point point) {
-	if (!LBItem::contains(point))
-		return false;
-
-	if (!_doHitTest)
-		return true;
-
-	return !_anim->transparentAt(point.x, point.y);
-}
-
-void LBAnimationItem::update() {
-	if (_loaded && _enabled && _globalEnabled && _running) {
-		bool wasDone = _anim->update();
-		if (wasDone) {
-			_running = false;
-			done(true);
-		}
-	}
-
-	LBItem::update();
-}
-
-bool LBAnimationItem::togglePlaying(bool playing, bool restart) {
-	if (playing) {
-		if (_loaded && _enabled && _globalEnabled) {
-			if (restart)
-				seek(1);
-			_running = true;
-			_anim->start();
-		}
-
-		return _running;
-	}
-
-	return LBItem::togglePlaying(playing, restart);
-}
-
-void LBAnimationItem::done(bool onlyNotify) {
-	if (!onlyNotify) {
-		_anim->stop();
-	}
-
-	LBItem::done(onlyNotify);
-}
-
-void LBAnimationItem::init() {
-	_anim = new LBAnimation(_vm, this, _resourceId);
-
-	LBItem::init();
-}
-
-void LBAnimationItem::stop() {
-	if (_running) {
-		_anim->stop();
-		seek(0xFFFF);
-	}
-
-	_running = false;
-
-	LBItem::stop();
-}
-
-void LBAnimationItem::seek(uint16 pos) {
-	_anim->seek(pos);
-}
-
-void LBAnimationItem::seekToTime(uint32 time) {
-	_anim->seekToTime(time);
-}
-
-void LBAnimationItem::startPhase(uint phase) {
-	if (phase == _phase)
-		seek(1);
-
-	LBItem::startPhase(phase);
-}
-
-void LBAnimationItem::draw() {
-	if (!_visible || !_globalVisible)
-		return;
-
-	_anim->draw();
-}
-
-LBItem *LBAnimationItem::createClone() {
-	LBAnimationItem *item = new LBAnimationItem(_vm, _page, _rect);
-	item->_anim = new LBAnimation(_vm, item, _resourceId);
-	return item;
-}
-
-LBMovieItem::LBMovieItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBMovieItem");
-}
-
-LBMovieItem::~LBMovieItem() {
-}
-
-void LBMovieItem::update() {
-	if (_playing) {
-		VideoEntryPtr video = _vm->_video->findVideo(_resourceId);
-		if (!video || video->endOfVideo())
-			done(true);
-	}
-
-	LBItem::update();
-}
-
-bool LBMovieItem::togglePlaying(bool playing, bool restart) {
-	if (playing) {
-		if ((_loaded && _enabled && _globalEnabled) || _phase == kLBPhaseNone) {
-			debug("toggled video for phase %d", _phase);
-			VideoEntryPtr video = _vm->_video->playMovie(_resourceId);
-			if (!video)
-				error("Failed to open tMOV %d", _resourceId);
-
-			video->moveTo(_rect.left, _rect.top);
-			return true;
-		}
-	}
-
-	return LBItem::togglePlaying(playing, restart);
-}
-
-LBItem *LBMovieItem::createClone() {
-	return new LBMovieItem(_vm, _page, _rect);
-}
-
-LBMiniGameItem::LBMiniGameItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBMiniGameItem");
-}
-
-LBMiniGameItem::~LBMiniGameItem() {
-}
-
-bool LBMiniGameItem::togglePlaying(bool playing, bool restart) {
-	// HACK: Since we don't support any of these hardcoded mini games yet,
-	// just skip to the most logical page. For optional minigames, this
-	// will return the player to the previous page. For mandatory minigames,
-	// this will send the player to the next page.
-
-	uint16 destPage = 0;
-	bool returnToMenu = false;
-
-	// Figure out what minigame we have and bring us back to a page where
-	// the player can continue
-	if (_desc == "Kitch")     // Green Eggs and Ham: Kitchen minigame
-		destPage = 4;
-	else if (_desc == "Eggs") // Green Eggs and Ham: Eggs minigame
-		destPage = 5;
-	else if (_desc == "Fall") // Green Eggs and Ham: Fall minigame
-		destPage = 13;
-	else if (_desc == "MagicWrite3") // Arthur's Reading Race: "Let Me Write" minigame (Page 3)
-		destPage = 3;
-	else if (_desc == "MagicWrite4") // Arthur's Reading Race: "Let Me Write" minigame (Page 4)
-		destPage = 4;
-	else if (_desc == "MagicSpy5") // Arthur's Reading Race: "I Spy" minigame (Page 5)
-		destPage = 5;
-	else if (_desc == "MagicSpy6") // Arthur's Reading Race: "I Spy" minigame (Page 6)
-		destPage = 6;
-	else if (_desc == "MagicWrite7") // Arthur's Reading Race: "Let Me Write" minigame (Page 7)
-		destPage = 7;
-	else if (_desc == "MagicSpy8") // Arthur's Reading Race: "I Spy" minigame (Page 8)
-		destPage = 8;
-	else if (_desc == "MagicRace") // Arthur's Reading Race: Race minigame
-		returnToMenu = true;
-	else
-		error("Unknown minigame '%s'", _desc.c_str());
-
-	GUI::MessageDialog dialog(Common::String::format("The '%s' minigame is not supported yet.", _desc.c_str()));
-	dialog.runModal();
-
-	// Go back to the menu if requested, otherwise go to the requested page
-	if (returnToMenu)
-		_vm->addNotifyEvent(NotifyEvent(kLBNotifyGoToControls, 1));
-	else
-		_vm->addNotifyEvent(NotifyEvent(kLBNotifyChangePage, destPage));
-
-	return false;
-}
-
-LBItem *LBMiniGameItem::createClone() {
-	error("can't clone LBMiniGameItem");
-}
-
-LBProxyItem::LBProxyItem(MohawkEngine_LivingBooks *vm, LBPage *page, Common::Rect rect) : LBItem(vm, page, rect) {
-	debug(3, "new LBProxyItem");
-
-	_page = nullptr;
-}
-
-LBProxyItem::~LBProxyItem() {
-	delete _page;
-}
-
-void LBProxyItem::load() {
-	if (_loaded)
-		return;
-
-	Common::String leftover;
-	Common::String filename = _vm->getFileNameFromConfig("Proxies", _desc.c_str(), leftover);
-	if (!leftover.empty())
-		error("LBProxyItem tried loading proxy '%s' but got leftover '%s'", _desc.c_str(), leftover.c_str());
-	uint16 baseId = 0;
-	for (uint i = 0; i < filename.size(); i++) {
-		if (filename[i] == ';') {
-			baseId = atoi(filename.c_str() + i + 1);
-			filename = Common::String(filename.c_str(), i);
-		}
-	}
-
-	debug(1, "LBProxyItem loading archive '%s' with id %d", filename.c_str(), baseId);
-	Archive *pageArchive = _vm->createArchive();
-	if (!tryOpenPage(pageArchive, filename))
-		error("failed to open archive '%s' (for proxy '%s')", filename.c_str(), _desc.c_str());
-	_page = new LBPage(_vm);
-	_page->open(pageArchive, baseId);
-
-	LBItem::load();
-}
-
-void LBProxyItem::unload() {
-	delete _page;
-	_page = nullptr;
-
-	LBItem::unload();
-}
-
-LBItem *LBProxyItem::createClone() {
-	return new LBProxyItem(_vm, _page, _rect);
 }
 
 } // End of namespace Mohawk
