+++ engines/mohawk/module.mk
@@ -8,9 +8,14 @@ MODULE_OBJS = \
 	graphics.o \
 	installer_archive.o \
 	livingbooks.o \
+	livingbooks_animation.o \
+	livingbooks_bookinfo.o \
 	livingbooks_code.o \
 	livingbooks_graphics.o \
+	livingbooks_itemtypes.o \
 	livingbooks_lbx.o \
+	livingbooks_menus.o \
+	livingbooks_page.o \
 	metaengine.o \
 	riven_metaengine.o \
 	myst_metaengine.o \
